Began preparations to develop a more realisitic nbody simulation.
The previous nbody simulation ignored gravitational and electric
forces acting on the particle. This is okay for simulations with
negiligable masses and charges where accuracy isn't particularly
important, but you will often want to create a model more in line
with reality. This feature will account for these previously
ignored forces.

To prepare for the new feature, I have modified the particle class
in the auxillary module to include charge and copied the nbody
script to use as a template for the updated simulation.